Add LetsEncrypt to all web virtualhosts
HTTPS vhosts are all disabled by default via $https parameters, which
allows for bootstrapping when the certificate is not yet available.

On the first run, LE will run using webroot mode and will retrieve a
cert, authenticating using files served by the HTTP vhosts.  On
subsequent runs, $https can be enabled and the HTTPS vhosts will be
configured with the cert.  The letsencrypt module sets up a cronjob for
automatic renewals via the webroot.

A single certificate with SANs is used for all virtual hosts, which even
avoids the need for SNI.