Wipe out old vagrant dirs
When upgrading from 1.4.2 to 2.0.2 vagrant wants to upgrade some
internal data. This shows an interactive prompt which needs a TTY. We
don't care about our boxes so it's easier to just remove the directory.
We only do this if we see the old rackspace dummy box to make it
idempotent.