try harder to get a node from cico and fail properly
sometimes, cico returns

    {"message": "Requested servers successfully", "hosts": {}, "ssid": "012345679"}

with the hosts list being empty. This makes the data useless for further
processing. Try to get a list of hosts from cico again and if that
returns no hosts too, properly fail inside the playbook, instead of
writing an empty inventory and then failing in the next steps of the
pipeline.