only add job link for the currently running job
when a pipeline runs multiple steps in parallel on the same node, each
step gets an unique workspace (jobname@XXX for all but the first run).
this means that when we write the remote job details to the workspace,
we might not overwrite any pre-existing files in there.

this leads to jobs with multiple links like seen in [1], which is
confusing.
let's be more explicit and only add one link at at time to the
description.

[1] https://ci.theforeman.org/job/katello-nightly-rpm-pipeline/659/