only set a long cache time for files, not indexes
Before this change, we would set Cache-Control for *ALL* of
downloads.theforeman.org, leading to the CDN caching directory indexes
which is confusing during release time when you push new files but don't
see them.

Intead, let's just set Cache-Control on the actual files we want to
serve from the CDN for a long time.