Remove the reverse dependency on the installer
We added this dependency when the installer was split over 2
repositories. This was intended to keep them closer in sync.

Now that it's only in the foreman repository, this isn't really needed
anymore.

The effect is that it really does become a nightly again, which will
reduce the spam on Discourse. That was increased because we built an
installer for every merge to foreman-packaging.