Add runCicoJob and runCicoJobsInParallel functions
These make it easy to run one or more cico jobs. This also changes the
model where we run each cico job on a different node to running it all
on a single node in parallel. This can be done because it's only polling
for a result and not actual compute.