Add a foreman-installer-develop-test job
This works towards the pattern of having a -test job to run tests and
then a -release job to actually release it.

An additional benefit is that we will stop building an installer for
every merged PR to foreman-packaging.

It is not complete since the tests with the actual modules, as we do in
Travis, are not ran. Those wouldn't be correct because there is a timing
issue. The -test job runs, then the -release job builds a tarbal from
potentially different sources.