Move to a source and package release model
This changes the idea of a 'test' job that is run on PR merge to
a source release job. The source release job aims to test the code
and then create a source release if that tested code is green.

The previous release job is renamed to be a package-release job to
better specify that its focused only on releasing a package.

The overall idea is to split source generation and release from
package creation given that source can be an input to multiple places.