Refactor plugins pipeline
This refactors the pipeline to be driven by configuration and allows
changing the versions by just changing the versions hash at the top. The
end result is still readable for end users in the same way. It may be
even more readable by printing the exception.

It also allows defining multiple distributions which will be useful when
we add EL8 support.