Move Katello react tests out of parallel steps
With the addition of functional tests that use a lot of asynchronous actions, we are seeing timeout issues for these tests in Jenkins but not in GH actions. Moving them out of the parallel steps should help to address these timeout issues. We have already increased the timeouts for jest, react-testing-lib, and nock.

The goal is to remove the react tests (and foreman npm install step) from the PR job and only keep it in the nightly/release job. This would allow us to rely on GH actions in PRs and speed up the jenkins job. I think it is a good idea to see the tests run successfully for a few days in the new non-parallel step first, so we can be sure there won't be issues during the release jobs.