collect debug from all machines that are named pipe-*
in #1547 we made the fallback "all", but also discussed that the whole
if/else logic might be obsolete.
"all" is not a perfect fallback as this also runs collect debug against
other hosts defined in the inventory -- especially the hypervisor we run
our tests on.
however we can use a "pipe-*" pattern to match all hosts belonging to
the pipeline and only fetch data from those.