explicitly set buildresult to ..
pbulder defaults to storing its build results in
/var/cache/pbuilder/result (and this is also hardcoded in [1]).
however, when using (p)debuild, you end up having your built debs in the
directory above the one you started the build in (= ..).
due to the fact that we explicitly configure BUILDRESULT ([2]) in our
pbuilderrc, it tries to copy files to /var/cache/pbuilder/result too,
which now fails as we don't call pdebuild with sudo anymore.

we can explicitly pass --buildresult .. to pdebuild, which will result in
a warning, as it tries to copy files to a place they are already present,
but otherwise this works well.

    I: file ../hello-dbgsym_2.10-2_amd64.deb is already in target, not copying.
    I: file ../hello_2.10-2_amd64.buildinfo is already in target, not copying.
    I: file ../hello_2.10-2_amd64.deb is already in target, not copying.
    I: file ../hello_2.10-2_amd64.changes is already in target, not copying.

[1] https://github.com/voxpupuli/puppet-pbuilder/blob/2ba75ff85ddabadf1532ea93991bbd1f58f24df9/manifests/init.pp#L65
[2] https://github.com/voxpupuli/puppet-pbuilder/blob/2ba75ff85ddabadf1532ea93991bbd1f58f24df9/templates/pbuilderrc.erb#L25