Ensure root SSH config provisions correctly
On a fresh installation the ssh module fails because /root/.ssh doesn't
exist and then authorized_keys can't be created. We actually don't need
this to be a file at all so we can stop managing it. If we would manage
SSH keys, the ssh_authorized_key resource type can create it if needed.