Add two playbooks to generate modules metadata and then add it to an existing repository
Module metadata needs to include the set of RPMs with full NEVRA
to allow dependency resolution. This assumes a base modules metadata
file and then analyzes a mashed repository to find all RPMs to add
dynamically to the metadata. The metadata is then added to the repository
via modifyrepo_c.