Fix mashing script
The finalloc parameter was not used and callers weren't passing it in.

It also writes out the comps file to a temporary file since it was being
passes its content rather than the path. It does now always create a
temporary file, even if no comps is passed. However, this overhead is
tiny compared to what comps is already doing and keeps code complexity
down.

Fixes: 3f4ee09d908dc7f947f451cad8c3ba68d18157d6