Rewrite ssh_keygen to ssh::keygen
This uses the modern Puppet function API and uses data types for
validation.

It also changes from a hash to positional arguments, which is probably a
worse API. Now it's not too bad because we don't really use the
additional arguments, but in the future it could be introduced again and
checked via a Struct data type.