Use --force with puppet generate types
When installing a new module it maintains the file timestamps from the
original tarball. puppet generate types uses the timestamp of when it
ran.

If you update a module to a release that was created before it ran
puppet types generate, it will never update the types. This can happen
with a module downgrade, or just when you regenerated types manually. By
using --force the timestamps are ignored.